refactor(cr): simlify useIsBacker hook
