+++ src/app/collective-rewards/leaderboard/BuildersLeaderBoardContent.tsx
@@ -1,12 +1,12 @@
+import { BuildersLeaderBoardTable } from '@/app/collective-rewards/leaderboard'
 import { useGetBuildersRewards } from '@/app/collective-rewards/rewards'
+import { Search, SearchContextProvider } from '@/app/collective-rewards/shared'
 import { useHandleErrors } from '@/app/collective-rewards/utils'
 import { LoadingSpinner } from '@/components/LoadingSpinner'
 import { Typography } from '@/components/Typography'
-import { Search, SearchContextProvider } from '@/app/collective-rewards/shared'
-import { getAddress } from 'viem'
+import { COINBASE_ADDRESS } from '@/lib/constants'
 import { tokenContracts } from '@/lib/contracts'
-import { BuildersLeaderBoardTable } from '@/app/collective-rewards/leaderboard'
-import { CoinBaseAddress } from '@/lib/constants'
+import { getAddress } from 'viem'
 
 const EmptyLeaderboard = () => (
   <div className="relative w-full py-6">
@@ -28,7 +28,7 @@ export const BuildersLeaderBoardContent = () => {
       symbol: 'RIF',
     },
     rbtc: {
-      address: CoinBaseAddress,
+      address: COINBASE_ADDRESS,
       symbol: 'RBTC',
     },
   }
