+++ src/app/collective-rewards/rewards/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useGetBackerRewardPerTokenPaid'
 export * from './useGetBackersRewardPercentage'
 export * from './useGetGaugesEvents'
 export * from './useGetGaugesNotifyReward'
@@ -6,4 +5,4 @@ export * from './useGetLastCycleDistribution'
 export * from './useGetNotifyRewardLogs'
 export * from './useGetPerTokenRewards'
 export * from './useGetRewardDistributionFinishedLogs'
-export * from './useIsBuilderOrBacker'
+export * from './useIsBacker'
